Add v6 tests (#674)
Replaces #528. I was just going off the features listed in the readme. There are a handful more we should test but this is a start.

Note that this includes a readme change for `state.href` that I believe was incorrectly documented. Let me know if I got that wrong.

Also note that we should delete the original `v6-tests` branch once this is merged. Leaving it for now for reference.