Fix inspect script (#654)
`discify` is a different module than intended. discify is an alternative
to disc that works as a plugin, but we were using it as a separate
command line program. This switches it out for disc. The disc command
line program is called `discify` which is probably where the confusion
came from.

Also swapped out unassertify and uglifyify for tinyify, which adds tree
shaking.